atualizando a página de detalhes
