+++ src/Pages/RestaurantPage/RestaurantPage.js
@@ -2,37 +2,43 @@ import React from "react";
 import { BASE_URL } from "../../Constants/URL";
 import useRequestData from "../../Hooks/useRequestData";
 import { useParams } from "react-router-dom";
-import Menu from "./RestaurantMenu";
+import { useProtectedPage } from '../../Hooks/useProtectedPage';
 
 const RestaurantPage = () => {
+  useProtectedPage();
   const params = useParams();
-  const restaurantes = useRequestData([], `${BASE_URL}/restaurants`);
+  const restaurante = useRequestData(
+    {},
+    `${BASE_URL}/restaurants/${params.id}`
+  );
 
-  console.log("restaurantes", restaurantes);
-
-  const renderRestaurant =
-    restaurantes.restaurants &&
-    restaurantes.restaurants.map((restaurante) => {
-      let cardRestaurant;
-      if (restaurante.id === params.id) {
-        cardRestaurant = (
-          <div key={restaurante.id}>
-            <img src={restaurante.logoUrl} alt={restaurante.name} />
-            <p>{restaurante.name}</p>
-            <p>{restaurante.category}</p>
-            <p>{restaurante.deliveryTime} min</p>
-            <p>Frete R${restaurante.shipping}</p>
-            <p>{restaurante.address}</p>
-          </div>
-        );
-      }
-      return cardRestaurant;
+  const listaProdutos =
+    restaurante.restaurant &&
+    restaurante.restaurant.products &&
+    restaurante.restaurant.products.map((produto) => {
+      return (
+        <div>
+          <img width="200" src={produto.photoUrl} alt={produto.name} />
+          <p>{produto.name}</p>
+          <p>{produto.description}</p>
+          <p>R$ {produto.price}</p>
+        </div>
+      );
     });
 
   return (
     <div>
-      {renderRestaurant}
-      <Menu />
+      {restaurante.restaurant && (
+        <div>
+          <img src={restaurante.restaurant.logoUrl} alt={restaurante.restaurant.name} />
+            <p>{restaurante.restaurant.name}</p>
+            <p>{restaurante.restaurant.category}</p>
+            <p>{restaurante.restaurant.deliveryTime} min</p>
+            <p>Frete R${restaurante.restaurant.shipping}</p>
+            <p>{restaurante.restaurant.address}</p>
+        </div>
+      )}
+      {listaProdutos}
     </div>
   );
 };
